+++ tests/testthat/examples/fastcpd_exponential_1.R
@@ -1,0 +1,9 @@
+set.seed(1)
+data <- matrix(c(
+  rexp(300, rate = 1),
+  rexp(400, rate = 0.1),
+  rexp(300, rate = 1)
+))
+system.time(result <- fastcpd.exponential(data))
+summary(result)
+plot(result)